Some restructuring, implemented example for native simple_ot
