+++ Cargo.toml
@@ -6,6 +6,10 @@ authors = ["Lennart Vogelsang <dev@vogelsangs.net>"]
 publish = false
 edition = "2018"
 
+[[bin]]
+name = "ot"
+path = "src/bin.rs"
+
 [lib]
 #TODO make lib work with web crate
 #crate-type = ["cdylib"]
@@ -15,17 +19,17 @@ path = "src/lib.rs"
 
 [dependencies]
 # general util
-failure = "0.1.2"
+failure = "0.1.3"
 byteorder = "1"
-generic-array = "0.9.0"
+generic-array = "0.12.0"
 # async & io
-futures-preview = "0.3.0-alpha.6"
-tokio-async-await = "0.1.0"
+futures-preview = "0.3.0-alpha.10"
+romio = "0.3.0-alpha.1"
 # crypto
-curve25519-dalek = {version = "0.19", features = ["avx2_backend", "std"], default-features = false}
+curve25519-dalek = {version = "0.21", features = ["avx2_backend", "std"], default-features = false}
 tiny-keccak = "1.4.2"
-ring = "0.12.1"
-rand = "0.5.0"
+ring = "0.13.5"
+rand = "0.6.1"
 
 # [[bin]]
 # name = "bin_test"
