+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![feature(async_await, await_macro, futures_api)]
-#[macro_use] 
-extern crate failure;
 
 pub mod base_ot;
 pub mod crypto;
 pub mod util;
+
+pub mod native;
