+++ src/native/aes.rs
@@ -1,13 +1,16 @@
-use super::{SymmetricCrypt};
+use crate::crypto::SymmetricCryptoProvider;
 use failure::{Fallible};
+use futures::future::{ready, Ready};
 use generic_array::{typenum::U32, GenericArray};
 use ring::aead::*;
 
+/// Implementation of the CryptoTrait for native environments. It wraps the AES
+/// functions of the ring library.
 #[derive(Default)]
 pub struct AesCryptoProvider();
 
-impl SymmetricCrypt<U32> for AesCryptoProvider {
-    fn encrypt(&mut self, key: &GenericArray<u8, U32>, mut data: Vec<u8>) -> Fallible<Vec<u8>> {
+impl AesCryptoProvider {
+    fn encrypt_inner(&mut self, key: &GenericArray<u8, U32>, mut data: Vec<u8>) -> Fallible<Vec<u8>> {
         // we can use a static 0 nonce here, because our always keys differ from message to message (TODO: prove that?!)
         let nonce: [u8; 12] = Default::default();
         let len = data.len() + AES_256_GCM.tag_len();
@@ -18,7 +21,7 @@ impl SymmetricCrypt<U32> for AesCryptoProvider {
         Ok(data)
     }
 
-    fn decrypt(&mut self, key: &GenericArray<u8, U32>, mut data: Vec<u8>) -> Fallible<(Vec<u8>)> {
+    fn decrypt_inner(&mut self, key: &GenericArray<u8, U32>, mut data: Vec<u8>) -> Fallible<Vec<u8>> {
         // we can use a static 0 nonce here, because our always keys differ from message to message (TODO: prove that?!)
         let nonce: [u8; 12] = Default::default();
         let opening_key = OpeningKey::new(&AES_256_GCM, key)?;
@@ -26,5 +29,15 @@ impl SymmetricCrypt<U32> for AesCryptoProvider {
         let len = data.len() - AES_256_GCM.tag_len();
         data.truncate(len);
         Ok(data)
+    }
+}
+
+impl SymmetricCryptoProvider<U32> for AesCryptoProvider {
+    type CryptReturn = Ready<Fallible<Vec<u8>>>;
+    fn encrypt(&mut self, key: &GenericArray<u8, U32>, data: Vec<u8>) -> Self::CryptReturn {
+        ready(self.encrypt_inner(key, data))
+    }
+    fn decrypt(&mut self, key: &GenericArray<u8, U32>, data: Vec<u8>) -> Self::CryptReturn {
+        ready(self.decrypt_inner(key, data))
     }
 }
