+++ src/util/mod.rs
@@ -1,6 +1,10 @@
 use byteorder::{BigEndian, ByteOrder};
 use failure::{Fallible};
 use futures::prelude::*;
+use rand::{thread_rng, Rng};
+use rand::distributions::Alphanumeric;
+
+// TODO: or specify length of data? More secure? Does this leak too much?
 
 pub async fn read_data<T: AsyncRead + AsyncReadExt>(conn: &mut T) -> Fallible<Vec<u8>> {
     let mut len_buf = [0; 8];
@@ -17,4 +21,14 @@ pub async fn write_data<'a, T: AsyncWrite + AsyncWriteExt>(conn: &'a mut T, data
     await!(conn.write_all(&mut len_buf))?;
     await!(conn.write_all(& data))?;
     Ok(())
+}
+
+pub fn generate_random_strings(n: usize, l: usize) -> Vec<String> {
+    let mut rng = thread_rng();
+    let mut values = Vec::with_capacity(n);
+    for _ in 0..n {
+        let s: String = rng.sample_iter(&Alphanumeric).take(l).collect();
+        values.push(s);
+    }
+    values
 }
